requirements.txt
Flask==1.1.2
numpy==1.22.2
opencv_python_headless==4.5.1.48
Pillow==9.2.0
requests==2.24.0
torch==1.7.1
torchvision==0.8.2
vgg_pytorch==0.3.0
